+++ MercadoPagoSDK/MercadoPagoSDK/RequestLayer/Checkout/CheckoutService.swift
@@ -1,4 +1,4 @@
-protocol PaymentService {
+protocol CheckoutService {
     func getInit(preferenceId: String?,
                  privateKey: String?,
                  body: Data?,
@@ -6,12 +6,12 @@ protocol PaymentService {
                  completion: @escaping (Swift.Result<PXInitDTO, PXError>) -> Void)
 }
 
-final class PaymentServiceImpl: PaymentService {
+final class CheckoutServiceImpl: CheckoutService {
     // MARK: - Private properties
-    private let service: Request<PaymentRequestInfos>
+    private let service: Request<CheckoutRequestInfos>
 
     // MARK: - Initialization
-    init(service: Request<PaymentRequestInfos> = Request<PaymentRequestInfos>()) {
+    init(service: Request<CheckoutRequestInfos> = Request<CheckoutRequestInfos>()) {
         self.service = service
     }
 
@@ -19,7 +19,7 @@ final class PaymentServiceImpl: PaymentService {
     func getInit(preferenceId: String?, privateKey: String?, body: Data?, headers: [String: String]?, completion: @escaping (Swift.Result<PXInitDTO, PXError>) -> Void) {
         service.requestObject(model: PXInitDTO.self, .getInit(preferenceId: preferenceId, privateKey: privateKey, body: body, headers: headers)) { apiResponse in
             switch apiResponse {
-            case .success(let payment): completion(.success(payment))
+            case .success(let mobileApiResult): completion(.success(mobileApiResult))
             case .failure: completion(.failure(PXError(domain: ApiDomain.GET_REMEDY,
                                                        code: ErrorTypes.NO_INTERNET_ERROR,
                                                        userInfo: [